# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  dto.CoffeeShopResponse:
    properties:
      address:
        type: string
      contacts:
        type: string
      id:
        type: string
      name:
        type: string
      rules:
        type: string
      welcome_message:
        type: string
    type: object
  dto.CreateCoffeeShopRequest:
    properties:
      address:
        type: string
      contacts:
        type: string
      name:
        type: string
      rules:
        type: string
      welcome_message:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.UpdateCoffeeShopRequest:
    properties:
      address:
        type: string
      contacts:
        type: string
      name:
        type: string
      rules:
        type: string
      welcome_message:
        type: string
    type: object
...
  dto.VerifyOTPRequest:
    properties:
      name:
        type: string
      otp:
        type: string
      phone:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /auth:
    post:
      consumes:
      - application/json
      description: Verify One-Time Password and authenticate user
      parameters:
      - description: OTP verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Verify OTP
      tags:
      - auth
  /auth/{phone}:
    get:
      description: Get One-Time Password for phone number
      parameters:
      - description: Phone number
        in: path
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get OTP
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user by invalidating refresh token
      parameters:
      - description: Logout request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - auth
  /auth/logout-everywhere:
    post:
      description: Logout user from all devices by invalidating all refresh tokens
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout Everywhere
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using a refresh token
      parameters:
      - description: Refresh token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Refresh Access Token
      tags:
      - auth
  /coffee-shops:
    get:
      description: Get a list of all coffee shops with optional pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CoffeeShopResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all coffee shops
      tags:
      - coffee-shops
    post:
      consumes:
      - application/json
      description: Create a new coffee shop with the provided information
      parameters:
      - description: Coffee shop information
        in: body
        name: coffee_shop
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCoffeeShopRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CoffeeShopResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new coffee shop
      tags:
      - coffee-shops
  /coffee-shops/{id}:
    delete:
      description: Delete a coffee shop by ID
      parameters:
      - description: Coffee Shop ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete coffee shop by ID
      tags:
      - coffee-shops
    get:
      description: Get coffee shop details by ID
      parameters:
      - description: Coffee Shop ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CoffeeShopResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get coffee shop by ID
      tags:
      - coffee-shops
    put:
      consumes:
      - application/json
      description: Update coffee shop details for the given ID
      parameters:
      - description: Coffee Shop ID
        in: path
        name: id
        required: true
        type: string
      - description: Coffee shop update information
        in: body
        name: coffee_shop
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCoffeeShopRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update coffee shop by ID
      tags:
      - coffee-shops
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"